implement view for other user
